Refactor tests to use repository instead of MVC, add @DirtiesContext to isolate side effects
* Switched from MockMvc to repository-based tests to streamline test execution and focus on database operations.
* Added @DirtiesContext annotation to ensure test isolation and prevent unintended side effects between tests.
