--- testing-modules/spring-testing-2/src/main/java/com/baeldung/dbteardown/CustomerController.java
@@ -1,39 +1,0 @@
-package com.baeldung.dbteardown;
-
-import java.util.List;
-import java.util.Optional;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class CustomerController {
-
-    private final CustomerRepository customerRepository;
-
-    public CustomerController(CustomerRepository customerRepository) {
-        this.customerRepository = customerRepository;
-    }
-
-    @PostMapping("/customers")
-    Customer addCustomer(@RequestBody Customer customer) {
-        return customerRepository.save(customer);
-    }
-
-    @GetMapping("/customers")
-    public List<Customer> getCustomers() {
-        return customerRepository.findAll();
-    }
-
-    @GetMapping("/customers/{name}")
-    public ResponseEntity<Customer> getCustomerByName(@PathVariable String name) {
-        Optional<Customer> customer = customerRepository.findByName(name);
-        return customer.map(ResponseEntity::ok)
-            .orElseGet(() -> ResponseEntity.notFound()
-                .build());
-    }
-}
